+++ finch-java-core/src/main/kotlin/com/tryfinch/api/models/PayStatement.kt
@@ -3075,6 +3075,8 @@ private constructor(
 
             @JvmField val DIRECT_DEPOSIT = of("direct_deposit")
 
+            @JvmField val OTHER = of("other")
+
             @JvmStatic fun of(value: String) = PaymentMethod(JsonField.of(value))
         }
 
@@ -3082,6 +3084,7 @@ private constructor(
         enum class Known {
             CHECK,
             DIRECT_DEPOSIT,
+            OTHER,
         }
 
         /**
@@ -3096,6 +3099,7 @@ private constructor(
         enum class Value {
             CHECK,
             DIRECT_DEPOSIT,
+            OTHER,
             /**
              * An enum member indicating that [PaymentMethod] was instantiated with an unknown
              * value.
@@ -3114,6 +3118,7 @@ private constructor(
             when (this) {
                 CHECK -> Value.CHECK
                 DIRECT_DEPOSIT -> Value.DIRECT_DEPOSIT
+                OTHER -> Value.OTHER
                 else -> Value._UNKNOWN
             }
 
@@ -3129,6 +3134,7 @@ private constructor(
             when (this) {
                 CHECK -> Known.CHECK
                 DIRECT_DEPOSIT -> Known.DIRECT_DEPOSIT
+                OTHER -> Known.OTHER
                 else -> throw FinchInvalidDataException("Unknown PaymentMethod: $value")
             }