+++ finch-java-core/src/main/kotlin/com/tryfinch/api/models/SandboxPaymentCreateParams.kt
@@ -3650,6 +3650,8 @@ private constructor(
 
                 @JvmField val DIRECT_DEPOSIT = of("direct_deposit")
 
+                @JvmField val OTHER = of("other")
+
                 @JvmStatic fun of(value: String) = PaymentMethod(JsonField.of(value))
             }
 
@@ -3657,6 +3659,7 @@ private constructor(
             enum class Known {
                 CHECK,
                 DIRECT_DEPOSIT,
+                OTHER,
             }
 
             /**
@@ -3671,6 +3674,7 @@ private constructor(
             enum class Value {
                 CHECK,
                 DIRECT_DEPOSIT,
+                OTHER,
                 /**
                  * An enum member indicating that [PaymentMethod] was instantiated with an unknown
                  * value.
@@ -3689,6 +3693,7 @@ private constructor(
                 when (this) {
                     CHECK -> Value.CHECK
                     DIRECT_DEPOSIT -> Value.DIRECT_DEPOSIT
+                    OTHER -> Value.OTHER
                     else -> Value._UNKNOWN
                 }
 
@@ -3705,6 +3710,7 @@ private constructor(
                 when (this) {
                     CHECK -> Known.CHECK
                     DIRECT_DEPOSIT -> Known.DIRECT_DEPOSIT
+                    OTHER -> Known.OTHER
                     else -> throw FinchInvalidDataException("Unknown PaymentMethod: $value")
                 }