+++ finch-java-core/src/main/kotlin/com/tryfinch/api/services/async/hris/company/payStatementItem/RuleServiceAsyncImpl.kt
@@ -5,6 +5,7 @@ package com.tryfinch.api.services.async.hris.company.payStatementItem
 import com.tryfinch.api.core.ClientOptions
 import com.tryfinch.api.core.JsonValue
 import com.tryfinch.api.core.RequestOptions
+import com.tryfinch.api.core.checkRequired
 import com.tryfinch.api.core.handlers.errorHandler
 import com.tryfinch.api.core.handlers.jsonHandler
 import com.tryfinch.api.core.handlers.withErrorHandler
@@ -25,6 +26,7 @@ import com.tryfinch.api.models.RuleCreateResponse
 import com.tryfinch.api.models.RuleDeleteResponse
 import com.tryfinch.api.models.RuleUpdateResponse
 import java.util.concurrent.CompletableFuture
+import kotlin.jvm.optionals.getOrNull
 
 class RuleServiceAsyncImpl internal constructor(private val clientOptions: ClientOptions) :
     RuleServiceAsync {
@@ -105,6 +107,9 @@ class RuleServiceAsyncImpl internal constructor(private val clientOptions: Clien
             params: HrisCompanyPayStatementItemRuleUpdateParams,
             requestOptions: RequestOptions,
         ): CompletableFuture<HttpResponseFor<RuleUpdateResponse>> {
+            // We check here instead of in the params builder because this can be specified
+            // positionally or in the params class.
+            checkRequired("ruleId", params.ruleId().getOrNull())
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.PUT)
@@ -157,6 +162,7 @@ class RuleServiceAsyncImpl internal constructor(private val clientOptions: Clien
                             .let {
                                 HrisCompanyPayStatementItemRuleListPageAsync.builder()
                                     .service(RuleServiceAsyncImpl(clientOptions))
+                                    .streamHandlerExecutor(clientOptions.streamHandlerExecutor)
                                     .params(params)
                                     .response(it)
                                     .build()
@@ -172,6 +178,9 @@ class RuleServiceAsyncImpl internal constructor(private val clientOptions: Clien
             params: HrisCompanyPayStatementItemRuleDeleteParams,
             requestOptions: RequestOptions,
         ): CompletableFuture<HttpResponseFor<RuleDeleteResponse>> {
+            // We check here instead of in the params builder because this can be specified
+            // positionally or in the params class.
+            checkRequired("ruleId", params.ruleId().getOrNull())
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.DELETE)