+++ finch-java-core/src/main/kotlin/com/tryfinch/api/services/async/sandbox/EmploymentServiceAsyncImpl.kt
@@ -5,6 +5,7 @@ package com.tryfinch.api.services.async.sandbox
 import com.tryfinch.api.core.ClientOptions
 import com.tryfinch.api.core.JsonValue
 import com.tryfinch.api.core.RequestOptions
+import com.tryfinch.api.core.checkRequired
 import com.tryfinch.api.core.handlers.errorHandler
 import com.tryfinch.api.core.handlers.jsonHandler
 import com.tryfinch.api.core.handlers.withErrorHandler
@@ -18,6 +19,7 @@ import com.tryfinch.api.core.prepareAsync
 import com.tryfinch.api.models.EmploymentUpdateResponse
 import com.tryfinch.api.models.SandboxEmploymentUpdateParams
 import java.util.concurrent.CompletableFuture
+import kotlin.jvm.optionals.getOrNull
 
 class EmploymentServiceAsyncImpl internal constructor(private val clientOptions: ClientOptions) :
     EmploymentServiceAsync {
@@ -48,6 +50,9 @@ class EmploymentServiceAsyncImpl internal constructor(private val clientOptions:
             params: SandboxEmploymentUpdateParams,
             requestOptions: RequestOptions,
         ): CompletableFuture<HttpResponseFor<EmploymentUpdateResponse>> {
+            // We check here instead of in the params builder because this can be specified
+            // positionally or in the params class.
+            checkRequired("individualId", params.individualId().getOrNull())
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.PUT)