+++ finch-java-core/src/main/kotlin/com/tryfinch/api/services/blocking/jobs/AutomatedService.kt
@@ -52,14 +52,35 @@ interface AutomatedService {
         create(JobAutomatedCreateParams.none(), requestOptions)
 
     /** Get an automated job by `job_id`. */
-    fun retrieve(params: JobAutomatedRetrieveParams): AutomatedAsyncJob =
-        retrieve(params, RequestOptions.none())
+    fun retrieve(jobId: String): AutomatedAsyncJob =
+        retrieve(jobId, JobAutomatedRetrieveParams.none())
+
+    /** @see [retrieve] */
+    fun retrieve(
+        jobId: String,
+        params: JobAutomatedRetrieveParams = JobAutomatedRetrieveParams.none(),
+        requestOptions: RequestOptions = RequestOptions.none(),
+    ): AutomatedAsyncJob = retrieve(params.toBuilder().jobId(jobId).build(), requestOptions)
+
+    /** @see [retrieve] */
+    fun retrieve(
+        jobId: String,
+        params: JobAutomatedRetrieveParams = JobAutomatedRetrieveParams.none(),
+    ): AutomatedAsyncJob = retrieve(jobId, params, RequestOptions.none())
 
     /** @see [retrieve] */
     fun retrieve(
         params: JobAutomatedRetrieveParams,
         requestOptions: RequestOptions = RequestOptions.none(),
     ): AutomatedAsyncJob
+
+    /** @see [retrieve] */
+    fun retrieve(params: JobAutomatedRetrieveParams): AutomatedAsyncJob =
+        retrieve(params, RequestOptions.none())
+
+    /** @see [retrieve] */
+    fun retrieve(jobId: String, requestOptions: RequestOptions): AutomatedAsyncJob =
+        retrieve(jobId, JobAutomatedRetrieveParams.none(), requestOptions)
 
     /**
      * Get all automated jobs. Automated jobs are completed by a machine. By default, jobs are
@@ -117,8 +138,24 @@ interface AutomatedService {
          * as [AutomatedService.retrieve].
          */
         @MustBeClosed
-        fun retrieve(params: JobAutomatedRetrieveParams): HttpResponseFor<AutomatedAsyncJob> =
-            retrieve(params, RequestOptions.none())
+        fun retrieve(jobId: String): HttpResponseFor<AutomatedAsyncJob> =
+            retrieve(jobId, JobAutomatedRetrieveParams.none())
+
+        /** @see [retrieve] */
+        @MustBeClosed
+        fun retrieve(
+            jobId: String,
+            params: JobAutomatedRetrieveParams = JobAutomatedRetrieveParams.none(),
+            requestOptions: RequestOptions = RequestOptions.none(),
+        ): HttpResponseFor<AutomatedAsyncJob> =
+            retrieve(params.toBuilder().jobId(jobId).build(), requestOptions)
+
+        /** @see [retrieve] */
+        @MustBeClosed
+        fun retrieve(
+            jobId: String,
+            params: JobAutomatedRetrieveParams = JobAutomatedRetrieveParams.none(),
+        ): HttpResponseFor<AutomatedAsyncJob> = retrieve(jobId, params, RequestOptions.none())
 
         /** @see [retrieve] */
         @MustBeClosed
@@ -126,6 +163,19 @@ interface AutomatedService {
             params: JobAutomatedRetrieveParams,
             requestOptions: RequestOptions = RequestOptions.none(),
         ): HttpResponseFor<AutomatedAsyncJob>
+
+        /** @see [retrieve] */
+        @MustBeClosed
+        fun retrieve(params: JobAutomatedRetrieveParams): HttpResponseFor<AutomatedAsyncJob> =
+            retrieve(params, RequestOptions.none())
+
+        /** @see [retrieve] */
+        @MustBeClosed
+        fun retrieve(
+            jobId: String,
+            requestOptions: RequestOptions,
+        ): HttpResponseFor<AutomatedAsyncJob> =
+            retrieve(jobId, JobAutomatedRetrieveParams.none(), requestOptions)
 
         /**
          * Returns a raw HTTP response for `get /jobs/automated`, but is otherwise the same as