+++ finch-java-core/src/main/kotlin/com/tryfinch/api/services/blocking/payroll/PayGroupServiceImpl.kt
@@ -5,6 +5,7 @@ package com.tryfinch.api.services.blocking.payroll
 import com.tryfinch.api.core.ClientOptions
 import com.tryfinch.api.core.JsonValue
 import com.tryfinch.api.core.RequestOptions
+import com.tryfinch.api.core.checkRequired
 import com.tryfinch.api.core.handlers.errorHandler
 import com.tryfinch.api.core.handlers.jsonHandler
 import com.tryfinch.api.core.handlers.withErrorHandler
@@ -19,6 +20,7 @@ import com.tryfinch.api.models.PayGroupRetrieveResponse
 import com.tryfinch.api.models.PayrollPayGroupListPage
 import com.tryfinch.api.models.PayrollPayGroupListParams
 import com.tryfinch.api.models.PayrollPayGroupRetrieveParams
+import kotlin.jvm.optionals.getOrNull
 
 class PayGroupServiceImpl internal constructor(private val clientOptions: ClientOptions) :
     PayGroupService {
@@ -56,6 +58,9 @@ class PayGroupServiceImpl internal constructor(private val clientOptions: Client
             params: PayrollPayGroupRetrieveParams,
             requestOptions: RequestOptions,
         ): HttpResponseFor<PayGroupRetrieveResponse> {
+            // We check here instead of in the params builder because this can be specified
+            // positionally or in the params class.
+            checkRequired("payGroupId", params.payGroupId().getOrNull())
             val request =
                 HttpRequest.builder()
                     .method(HttpMethod.GET)