+++ finch-java-core/src/test/kotlin/com/tryfinch/api/services/blocking/payroll/PayGroupServiceTest.kt
@@ -4,7 +4,6 @@ package com.tryfinch.api.services.blocking.payroll
 
 import com.tryfinch.api.TestServerExtension
 import com.tryfinch.api.client.okhttp.FinchOkHttpClient
-import com.tryfinch.api.models.PayrollPayGroupRetrieveParams
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 
@@ -20,10 +19,7 @@ internal class PayGroupServiceTest {
                 .build()
         val payGroupService = client.payroll().payGroups()
 
-        val payGroup =
-            payGroupService.retrieve(
-                PayrollPayGroupRetrieveParams.builder().payGroupId("pay_group_id").build()
-            )
+        val payGroup = payGroupService.retrieve("pay_group_id")
 
         payGroup.validate()
     }